+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - envs: Log action, observation and reward to "env" sub-dict of action
+- envs: Wrapper to act on the discrete-derivative of an action
+- envs: Wrapper to add the action to the observation vector
+- envs: Wrapper to low-pass filter action (modeling a delay)
 - envs: Wrapper to noisify actions
 - envs: Wrapper to noisify observations
 - spines: Label pi3hat spine log filenames with the robot's hostname