+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2023-12-01
+
 ### Added
 
 - PPO balancer: Convenience Makefile to manage the training directory
@@ -365,7 +367,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Starting this changelog.
 
-[unreleased]: https://github.com/upkie/upkie/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/upkie/upkie/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/upkie/upkie/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/upkie/upkie/compare/v1.5.0...v2.0.0
 [1.5.0]: https://github.com/upkie/upkie/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/upkie/upkie/compare/v1.3.4...v1.4.0